backend/requirements.txt
asgiref==3.7.2
# certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-debug-toolbar==4.1.0
django-filter==23.2
django-simple-history==3.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
Markdown==3.5

django-cors-headers==4.2.0

python-telegram-bot==20.6
aiogram==3.1.1

redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.4

Pillow==10.0.1

Werkzeug==3.0.0
pyOpenSSL==23.2.0
django-extensions==3.2.3
